bootstrap: fix conversion of subpackages; special-case uchar
* most (META) subpackages have a 'directory' property pointing to a
subdir containing resources, but some do not. Example: opam-lib. This
commit fixes the handling of such subpackages. BUILD.bazel files for
such no-dir subpkgs are installed in a (created) subdir named by the
subpackage, but the targets point to the original directory. The bug
was that targets were pointing to the subdir, which does not exist in
the original opam repo.
* uchar has a META file but it is empty (except for description and
version); the resources are part of the OCaml kernal/stdlib. So in
converting META files, a dep on 'uchar' (in "requires" prop) are
ignored. For now; we may need to do something better to handle
different versions.